Refactor tmi to seperate file
